backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0

sqlalchemy==2.0.36
psycopg==3.2.3
python-dotenv==1.0.1

redis==5.0.8
httpx==0.28.1

python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]
